pyproject.toml
[project]
name = "gemma4-31b"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "huggingface-hub>=1.9.2",
    "openpyxl>=3.1.0",
    "pillow>=10.0.0",
    "pillow-heif>=0.16.0",
    "pydantic-settings>=2.6.0",
    "pypdf>=5.0.0",
    "python-docx>=1.2.0",
    "uvicorn[standard]>=0.30.0",
    "xlrd>=2.0.1",
]
